@@ -4,7 +4,7 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.11.0 <3.0.0"
+  sdk: ">=3.1.2 <4.0.0"
 
 dependencies:
   flutter:
