@@ -4,14 +4,14 @@ import 'package:flutter/material.dart';
 
 class PieceView extends StatelessWidget {
   static final _pieceColors = [
-    [Colors.grey[800], Colors.black],
-    [Colors.white, Colors.grey[400]]
+    [Colors.grey[800]!, Colors.black],
+    [Colors.white, Colors.grey[400]!]
   ];
 
   final Color _textColor;
   final List<Color> _gradeColors;
   final PieceLayout layout;
-  PieceView({@required this.layout})
+  PieceView({required this.layout})
       : _gradeColors = _pieceColors[layout.pieceID.sign == -1 ? 0 : 1],
         _textColor = layout.pieceID.sign == -1 ? Colors.white : Colors.black;
 
@@ -26,8 +26,11 @@ class PieceView extends StatelessWidget {
       : Container(
           decoration: BoxDecoration(
             shape: BoxShape.circle,
-            gradient: LinearGradient(begin: Alignment.topLeft, colors: _gradeColors),
-            border: Border.all(color: layout.highlight ? Colors.yellow : Colors.black, width: layout.highlight ? 2 : 1),
+            gradient:
+                LinearGradient(begin: Alignment.topLeft, colors: _gradeColors),
+            border: Border.all(
+                color: layout.highlight ? Colors.yellow : Colors.black,
+                width: layout.highlight ? 2 : 1),
           ),
           child: Center(
             child: FractionallySizedBox(
@@ -36,7 +39,9 @@ class PieceView extends StatelessWidget {
                 child: Container(
                   decoration: BoxDecoration(
                     shape: BoxShape.circle,
-                    gradient: LinearGradient(begin: Alignment.topLeft, colors: [_gradeColors[1], _gradeColors[0]]),
+                    gradient: LinearGradient(
+                        begin: Alignment.topLeft,
+                        colors: [_gradeColors[1], _gradeColors[0]]),
                   ),
                   child: Center(
                     child: Text(
@@ -59,28 +64,31 @@ class PieceLayout {
 
   final int pipNo;
   final int pieceID;
-  final Offset offset;
+  final Offset? offset;
   final String label;
   final bool highlight;
   final bool edge;
 
   PieceLayout({
-    @required this.pipNo,
-    @required this.pieceID,
-    @required this.offset,
-    @required this.label,
+    required this.pipNo,
+    required this.pieceID,
+    required this.offset,
+    required this.label,
     this.highlight = false,
     this.edge = false,
   });
 
   Size get size => edge ? _edgeSize : _pieceSize;
-  Rect get rect => offset & size;
-  PieceLayout get animated => PieceLayout(pieceID: pieceID, offset: offset, label: '', pipNo: pipNo);
+  Rect get rect => offset! & size;
+  PieceLayout get animated =>
+      PieceLayout(pieceID: pieceID, offset: offset, label: '', pipNo: pipNo);
 
   @override
-  String toString() => 'layout(id=$pieceID, pipNo=$pipNo, label=$label, rect=$rect, highlight=$highlight)';
+  String toString() =>
+      'layout(id=$pieceID, pipNo=$pipNo, label=$label, rect=$rect, highlight=$highlight)';
 
-  static Iterable<PieceLayout> getLayouts(List<List<int>> board, [List<int> pipNosToHighlight]) sync* {
+  static Iterable<PieceLayout> getLayouts(List<List<int>> board,
+      [List<int?>? pipNosToHighlight]) sync* {
     assert(board.length == 26);
     assert(_pieceSize.width == _pieceSize.height);
 
@@ -99,7 +107,9 @@ class PieceLayout {
 
         for (var h = 0; h != pieceCount; ++h) {
           // if there's more than 5, the last one gets a label w/ the total number of pieces in the stack
-          final label = pieceCount > 5 && (h + 1) == pieceCount ? pieceCount.toString() : '';
+          final label = pieceCount > 5 && (h + 1) == pieceCount
+              ? pieceCount.toString()
+              : '';
           final dy = _offset.dy * min(4, h);
           final highlight = highlightedPiecePip && h == pieceCount - 1;
           final pieceID = pip[h];
@@ -107,19 +117,35 @@ class PieceLayout {
           if (pipNo >= 1 && pipNo <= 6) {
             // bottom right
             yield PieceLayout(
-                pipNo: pipNo, pieceID: pieceID, offset: Offset(468 - dx, 371 - dy), label: label, highlight: highlight);
+                pipNo: pipNo,
+                pieceID: pieceID,
+                offset: Offset(468 - dx, 371 - dy),
+                label: label,
+                highlight: highlight);
           } else if (pipNo >= 7 && pipNo <= 12) {
             // bottom left
             yield PieceLayout(
-                pipNo: pipNo, pieceID: pieceID, offset: Offset(204 - dx, 371 - dy), label: label, highlight: highlight);
+                pipNo: pipNo,
+                pieceID: pieceID,
+                offset: Offset(204 - dx, 371 - dy),
+                label: label,
+                highlight: highlight);
           } else if (pipNo >= 13 && pipNo <= 18) {
             // top left
             yield PieceLayout(
-                pipNo: pipNo, pieceID: pieceID, offset: Offset(24 + dx, 21 + dy), label: label, highlight: highlight);
+                pipNo: pipNo,
+                pieceID: pieceID,
+                offset: Offset(24 + dx, 21 + dy),
+                label: label,
+                highlight: highlight);
           } else if (pipNo >= 19 && pipNo <= 24) {
             // top right
             yield PieceLayout(
-                pipNo: pipNo, pieceID: pieceID, offset: Offset(288 + dx, 21 + dy), label: label, highlight: highlight);
+                pipNo: pipNo,
+                pieceID: pieceID,
+                offset: Offset(288 + dx, 21 + dy),
+                label: label,
+                highlight: highlight);
           } else {
             assert(false);
           }
@@ -131,27 +157,46 @@ class PieceLayout {
     for (final player in GammonPlayer.values) {
       final barPipNo = GammonRules.barPipNoFor(player);
       final highlightedPiecePip = pipNosToHighlight.contains(barPipNo);
-      final pieces = board[barPipNo].where((p) => GammonRules.playerFor(p) == player).toList();
+      final pieces = board[barPipNo]
+          .where((p) => GammonRules.playerFor(p) == player)
+          .toList();
       final pieceCount = pieces.length;
       for (var i = 0; i != pieceCount; ++i) {
         final pieceID = pieces[i];
-        final label = (i + 1) == pieceCount && pieceCount > 3 ? pieceCount.toString() : '';
-        final top = pieceID.sign == -1 ? 254.0 + _offset.dy * min(i, 2) : 138.0 - _offset.dy * min(i, 2);
+        final label = (i + 1) == pieceCount && pieceCount > 3
+            ? pieceCount.toString()
+            : '';
+        final top = pieceID.sign == -1
+            ? 254.0 + _offset.dy * min(i, 2)
+            : 138.0 - _offset.dy * min(i, 2);
         final highlight = highlightedPiecePip && i == 0;
         yield PieceLayout(
-            pipNo: barPipNo, pieceID: pieceID, offset: Offset(246, top), label: label, highlight: highlight);
+            pipNo: barPipNo,
+            pieceID: pieceID,
+            offset: Offset(246, top),
+            label: label,
+            highlight: highlight);
       }
     }
 
     // draw the pieces born off
     for (final player in GammonPlayer.values) {
       final offPipNo = GammonRules.offPipNoFor(player);
-      final pieces = board[offPipNo].where((p) => GammonRules.playerFor(p) == player).toList();
+      final pieces = board[offPipNo]
+          .where((p) => GammonRules.playerFor(p) == player)
+          .toList();
       final pieceCount = pieces.length;
       for (var i = 0; i != pieceCount; ++i) {
         final pieceID = pieces[i];
-        final top = pieceID.sign == -1 ? 386.0 - (_edgeSize.height + 1) * i : 22.0 + (_edgeSize.height + 1) * i;
-        yield PieceLayout(pipNo: offPipNo, pieceID: pieceID, offset: Offset(520, top), label: '', edge: true);
+        final top = pieceID.sign == -1
+            ? 386.0 - (_edgeSize.height + 1) * i
+            : 22.0 + (_edgeSize.height + 1) * i;
+        yield PieceLayout(
+            pipNo: offPipNo,
+            pieceID: pieceID,
+            offset: Offset(520, top),
+            label: '',
+            edge: true);
       }
     }
   }
